Added the ability to load config from the user's home directory
This changes the way that config loading works, so that in addition to
checking in the current directory for a config file, scss-lint will also
check in the user's home directory, and load .scss-lint.yml from there
instead if it exists.

This changes the precedence of config loading to:

1. Command line argument
2. .scss-lint.yml in current directory
3. .scss-lint.yml in home directory
4. Config file packaged in the gem

In all cases only the config file loaded is used - config from multiple
files is not merged together.

Fixes #436