Isolate CLI specs from repo directory
It was discovered that the presence of an `.scss-lint.yml` file in the
repo directory would break specs, since that file would be loaded
unexpectedly.

Fix this by running the CLI specs in an isolated environment. This
required us to change the `Config::USER_FILE` constant to a class method
so it could be evaluated dynamically, since the home directory could
change for different tests.