Remove XML formatter
As was mentioned in:

  https://github.com/brigade/scss-lint/pull/372#issuecomment-76219850

...this formatter doesn't appear to be used, and is superseded by the
`Checkstyle` plugin.

Removing it allows us to remove a dependency on Nokogiri, which
significantly speeds up test runs in Travis since it does not need to be
installed.

It also removes a source of confusion, as some developers mistake it for
the `Checkstyle` plugin at first.