Allow plugin configurations to override defaults
The previous implementation prevented the user from defining custom
plugin configuration that overrode the default configuration. This was
due to `Config#load_plugins` being called after the default
configuration had already been merged. Since the plugins allow their
configuration to be overridden by the local config, the default
configuration would always trump the plugin configuration.

Fix this by calling `load_plugins` in the `Config#load` helper. This
isn't really ideal as it tightly couples plugin configuration loading to
be a responsibility of the `Config` class itself, but to change this
would require some serious rethinking.

In the interest of shipping something that could benefit users now (and
which in theory we could still refactor to fix at some point in the
future), implement this less-than-ideal solution.