Fix `PropertyUnits` for multiple unit-line values
I noticed that this linter didn't correctly check all of the unit-like
parts of the value, which can happen pretty easily if you are using
shorthand values. For example:

  margin: 1px 2px 3pix 4px;

would have not been caught previously whereas:

  margin: 1pix 2px 3px 4px;

would have. I fixed this by finding all of the unit-like parts of the
value and checking the units on each of them. I moved the regex up to a
constant to avoid Rubocop's too many lines in method warning.