Allow passing file contents via standard input
Various third-party tools want the ability to lint a file passed via
standard input. A quirk of this feature is that there's no way to
specify a file name in this way, so we need to expose a flag that can be
specified in addition so that the appropriate linter configuration can
be applied based on the path.