Don't visit CommentNode children by default
A variety of linters fail on interpolation in /*...*/ comments due to
bugs in the source ranges returned by the Sass parser. Since it is
unlikely that developers care about linting interpolation in these
comments (all bug reports I've seen have not actually intended to
use the output for anything), disabling them by default should prevent
this class of problem.