Extract list of pseudo elements to data txt file
RuboCop complained about the number of lines in this class, so I decided
to extract the list of pseudo elements out into a txt file in the data
directory.

While I was at it, I added some copy to the readme similar to what we
have for the PropertySpelling documentation about submitting an issue to
add pseudo elements to this list.