Enable universal newline when reading scss files
SCSS-Lint seems to choke on a file that uses windows-style newlines with
carriage returns. (To type a windows-style newline with carriage return
in Vim, press <Ctrl>+v<Ctrl>+m in insert mode. This will show up as ^M.)

> SCSSLint::Linter::Indentation raised unexpected error linting file
> ./path/to/file.scss: 'undefined method `[]' for nil:NilClass'

To avoid this, we can encode the file contents using String#encode's
:universal_newline option, which

> Replaces CRLF (“rn”) and CR (“r”) with LF (“n”) if value is true.

http://ruby-doc.org/core-2.2.0/String.html#method-i-encode

I thought that maybe we wouldn't need the force_encoding anymore after
this change, but that was not the case since the test that was added
with that change started failing once I removed the force_encoding.

Fixes #700.