Fix linter_options to actually allow setting options
The way it's currently phrased, each time you call linter_options it
returns a new hash, because it's merging a hash literal with the
existing options. Thus you can never change the options.

Note: This doesn't take care of setting the value of
`@options['linters'][linter_name]` to the new hash if a new hash was
returned by fetch, but neither does the existing code so I'm not sure if
that's also desired.