Support Sass 3.5
In Sass 3.5, `node.value` returns an Array instead of a String.
This could mean that we could expect multiple strings. With this PR,
SCSS-Lint will check the first string returned.

Also, Sass 3.5.3 fixes a couple of issues which previously need a monkey
patch in SCSS-Lint. This PR uses '~> 3.5.3' as the minimum required
version of Sass

Fix: brigade/scss-lint#877

Ref: sass/sass#1799, sass/sass#2394